fix(invoker): check for a valid device name in config
Check that the device we are about to perform a command on
exists in the configuration before initiation a command. If no
valid device is found a message is printed to the screen and
a return code of 0 is returned (for that command). If multiple
devices are passed in a list, only the non-existant devices are
skipped.

Change-Id: I0e10edd21e9002c7320b636008f7e6edeb5184d1
Signed-off-by: dstott <devyn.stott@intel.com>