feat(cli): remove print functions from command invoker
Moved all print statements out of the invoker and into the cli.
We don't know if in the future the invoker will have a terminal
to print to. So all print statement should be logged instead, I
also made sure that every command invoked by the invoker
returns a command result. This was done so that these messages
could be printed nicely later on.

I also added device name to the commands. And prettied up the
CLI interface.

Successfull commands print to stdout, failed command print to
stderr.

While testing I also discovered and fixed a blug with SLURM.
A bug existed were mutiple partitions in SLURM would cause
control to inacurately state the a device was not configured
 in SLURM. This fixes that. Added some tests to handle it too.

Renamed classes in the CLI to better reflect there role and
avoid confusion. I added some additional tests as well.

Change-Id: Ifd47bf656de5360d66494d27b203d6e066aae8c3
Signed-off-by: dstott <devyn.stott@intel.com>